usm: Disable map no_prealloc if kernel is prior to 6.1
On kernel 6.1, bpf memory allocator was introduced, it takes care of caching allocations, without it, we can lead the system
into a bad performance.
