+++ pkg/network/config/usm_config_test.go
@@ -109,31 +109,6 @@ func TestUSMDataChannelSize(t *testing.T) {
 	})
 }
 
-func TestDisableMapPreallocation(t *testing.T) {
-	t.Run("via yaml", func(t *testing.T) {
-		mockSystemProbe := mock.NewSystemProbe(t)
-		mockSystemProbe.SetWithoutSource("service_monitoring_config.disable_map_preallocation", false)
-		cfg := New()
-
-		assert.False(t, cfg.DisableMapPreallocation)
-	})
-
-	t.Run("via ENV variable", func(t *testing.T) {
-		mock.NewSystemProbe(t)
-		t.Setenv("DD_SERVICE_MONITORING_CONFIG_DISABLE_MAP_PREALLOCATION", "false")
-		cfg := New()
-
-		assert.False(t, cfg.DisableMapPreallocation)
-	})
-
-	t.Run("default value", func(t *testing.T) {
-		mock.NewSystemProbe(t)
-		cfg := New()
-
-		assert.True(t, cfg.DisableMapPreallocation)
-	})
-}
-
 func TestMaxUSMConcurrentRequests(t *testing.T) {
 	t.Run("default value", func(t *testing.T) {
 		mock.NewSystemProbe(t)
