+++ pkg/system-probe/config/config_darwin.go
@@ -9,6 +9,11 @@ package config
 
 import "github.com/DataDog/datadog-agent/pkg/config/model"
 
+// eBPFMapPreallocationSupported returns false on non linux_bpf systems.
+func eBPFMapPreallocationSupported() bool {
+	return false
+}
+
 // ProcessEventDataStreamSupported returns true if process event data stream is supported
 func ProcessEventDataStreamSupported() bool {
 	return false
