+++ pkg/system-probe/config/config_linux.go
@@ -11,6 +11,11 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/config/model"
 )
 
+// eBPFMapPreallocationSupported returns false on non linux_bpf systems.
+func eBPFMapPreallocationSupported() bool {
+	return false
+}
+
 // ProcessEventDataStreamSupported returns true if process event data stream is supported
 func ProcessEventDataStreamSupported() bool {
 	return false
