+++ pkg/system-probe/config/config_unsupported.go
@@ -22,6 +22,11 @@ func ValidateSocketAddress(sockPath string) error {
 	return errors.New("system-probe unsupported")
 }
 
+// eBPFMapPreallocationSupported returns false on non linux_bpf systems.
+func eBPFMapPreallocationSupported() bool {
+	return false
+}
+
 // ProcessEventDataStreamSupported returns true if process event data stream is supported
 func ProcessEventDataStreamSupported() bool {
 	return false
